requirements.txt
# VentureScope Backend – Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.2
email-validator>=2.0.0
pydantic-settings==2.6.1

# OAuth 2.0
authlib==1.3.2

# Utils
python-dotenv==1.0.1

# AWS S3
boto3>=1.34.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.28.1
faker==33.1.0
pytest-cov==6.0.0
pgvector==0.2.5
openai
langchain
langchain-community
langchain-openai
# numpy<2  # Pinned since some ML packages have issues with numpy 2 out of the box in older environments
